mover scheduled_event_data_factory from conftest.py to tests/app/test_scheduled.py
