+++ tests/app/test_scheduled.py
@@ -4,9 +4,41 @@ from pytest_mock import MockerFixture
 from corva.app.base import BaseApp
 from corva.app.scheduled import ScheduledApp
 from corva.app.utils.context import ScheduledContext
+from corva.event.data.scheduled import ScheduledEventData
 from corva.event.event import Event
 from corva.event.loader.scheduled import ScheduledLoader
 from tests.conftest import ComparableException
+
+
+@pytest.fixture(scope='session')
+def scheduled_event_data_factory():
+    def _scheduled_event_data_factory(**kwargs):
+        for key, val in dict(
+             cron_string=str(),
+             environment=str(),
+             app=int(),
+             app_key=str(),
+             app_version=None,
+             app_connection_id=int(),
+             app_stream_id=int(),
+             source_type=str(),
+             company=int(),
+             provider=str(),
+             schedule=int(),
+             interval=int(),
+             schedule_start=int(),
+             schedule_end=int(),
+             asset_id=int(),
+             asset_name=str(),
+             asset_type=str(),
+             timezone=str(),
+             log_type=str()
+        ).items():
+            kwargs.setdefault(key, val)
+
+        return ScheduledEventData(**kwargs)
+
+    return _scheduled_event_data_factory
 
 
 @pytest.fixture(scope='function')
